OWL-4: Fixed a bug detected by the TestReplacePersonRequestSuccess test case
The handler didn't ignore the id provided in the request payload. It created a new record when the
id specified in the payload and the one specified in the request URL were different. As a result of
this bug fix, the handler will always ignore the id specified in the request payload.